requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==0.7.16
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.8
asttokens==3.0.0
async-timeout==4.0.3
asyncssh==2.19.0
atomicwrites==1.4.1
attrs==24.3.0
audioread==3.0.1
autopep8==2.0.4
babel==2.16.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.10.0
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cookiecutter==2.6.0
cryptography==44.0.0
ctransformers==0.2.27
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
diff-match-patch==20241021
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring-to-markdown==0.15
docutils==0.21.2
durationpy==0.9
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.1
flatbuffers==24.12.23
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.3.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
ImageHash==4.3.1
imagesize==1.4.1
importlib-metadata==6.11.0
importlib_resources==6.5.2
inflection==0.5.1
InstructorEmbedding==1.0.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.18.1
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.5
jiter==0.8.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.6.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.2.16
langchain-chroma==0.1.3
langchain-community==0.3.14
langchain-core==0.2.43
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.0
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multimethod==1.12
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.21.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
numpydoc==1.8.0
oauthlib==3.2.2
ollama==0.4.5
onnxruntime==1.19.2
openai==0.28.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandas-profiling==3.6.6
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.8
pexpect==4.9.0
phik==0.12.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pooch==1.8.2
posthog==3.7.5
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==3.2.0
PyGithub==2.5.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.3
pylint-venv==3.0.4
pyls-spyder==0.4.0
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
pythainlp==5.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.2
pyuca==1.2
PyWavelets==1.6.0
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rope==1.13.0
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.3.1
shap==0.46.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.13.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.3
spyder-kernels==3.0.2
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.29.0
streamlit-chat==0.1.1
streamlit-elements==0.1.0
streamlit-pandas-profiling==0.1.3
streamlit-shap==1.0.2
streamlit_mic_recorder==0.0.8
superqt==0.7.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.5.0
three-merge==0.1.1
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
visions==0.7.6
watchdog==4.0.2
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
whatthepatch==1.0.7
wordcloud==1.9.4
wrapt==1.17.0
xgboost==2.1.3
yapf==0.43.0
yarl==1.18.3
ydata-profiling==4.12.1
yfinance==0.2.51
zipp==3.21.0
